+++ ElvUI_LuckyoneUI/Core/Credits.lua
@@ -1,7 +1,6 @@
 local L1UI, E, L, V, P, G = unpack(select(2, ...))
 
-local sort = sort
-local pairs = pairs
+local pairs, sort = pairs, sort
 local tinsert = table.insert
 local tconcat = table.concat