+++ ElvUI_LuckyoneUI/Core/Layout.lua
@@ -42,7 +42,7 @@ function L1UI:SetupLayout(layout)
 	E.db["general"]["backdropcolor"]["b"] = 0.10196078431373
 	E.db["general"]["backdropcolor"]["g"] = 0.10196078431373
 	E.db["general"]["backdropcolor"]["r"] = 0.10196078431373
-	E.db["general"]["backdropfadecolor"]["a"] = 0.75
+	E.db["general"]["backdropfadecolor"]["a"] = 0.80
 	E.db["general"]["backdropfadecolor"]["b"] = 0.058823529411765
 	E.db["general"]["backdropfadecolor"]["g"] = 0.058823529411765
 	E.db["general"]["backdropfadecolor"]["r"] = 0.058823529411765
@@ -348,7 +348,7 @@ function L1UI:SetupLayout(layout)
 	E.db["chat"]["lfgIcons"] = false
 	E.db["chat"]["maxLines"] = 200
 	E.db["chat"]["numScrollMessages"] = 2
-	E.db["chat"]["panelColor"]["a"] = 0.7531320899725
+	E.db["chat"]["panelColor"]["a"] = 0.80
 	E.db["chat"]["panelColor"]["b"] = 0.058823529411765
 	E.db["chat"]["panelColor"]["g"] = 0.058823529411765
 	E.db["chat"]["panelColor"]["r"] = 0.058823529411765
@@ -419,8 +419,6 @@ function L1UI:SetupLayout(layout)
 	-- Tooltip
 	E.db["tooltip"]["alwaysShowRealm"] = true
 	E.db["tooltip"]["colorAlpha"] = 0.75
-	E.db["tooltip"]["dungeonScore"] = false
-	E.db["tooltip"]["dungeonScoreColor"] = false
 	E.db["tooltip"]["font"] = "Expressway"
 	E.db["tooltip"]["fontOutline"] = "OUTLINE"
 	E.db["tooltip"]["headerFontSize"] = 11
@@ -430,11 +428,16 @@ function L1UI:SetupLayout(layout)
 	E.db["tooltip"]["healthBar"]["statusPosition"] = "TOP"
 	E.db["tooltip"]["itemCount"] = "NONE"
 	E.db["tooltip"]["itemQuality"] = true
-	E.db["tooltip"]["mythicDataEnable"] = false
 	E.db["tooltip"]["showElvUIUsers"] = true
 	E.db["tooltip"]["showMount"] = false
 	E.db["tooltip"]["smallTextFontSize"] = 11
 	E.db["tooltip"]["textFontSize"] = 11
+
+	if E.Retail then
+		E.db["tooltip"]["inspectDataEnable"] = false
+		E.db["tooltip"]["mythicDataEnable"] = false
+		E.db["tooltip"]["role"] = false
+	end
 
 	-- Custom Text: Arena
 	E.db["unitframe"]["units"]["arena"]["customTexts"] = E.db["unitframe"]["units"]["arena"]["customTexts"] or {}
@@ -847,7 +850,7 @@ function L1UI:SetupLayout(layout)
 		end
 
 		-- UnitFrames DPS/TANK
-		E.db["unitframe"]["colors"]["castbar_backdrop"]["a"] = 0.75
+		E.db["unitframe"]["colors"]["castbar_backdrop"]["a"] = 0.80
 		E.db["unitframe"]["colors"]["castbar_backdrop"]["b"] = 0.058823529411765
 		E.db["unitframe"]["colors"]["castbar_backdrop"]["g"] = 0.058823529411765
 		E.db["unitframe"]["colors"]["castbar_backdrop"]["r"] = 0.058823529411765
@@ -1503,7 +1506,7 @@ function L1UI:SetupLayout(layout)
 		end
 
 		-- UnitFrames Healing
-		E.db["unitframe"]["colors"]["castbar_backdrop"]["a"] = 0.75
+		E.db["unitframe"]["colors"]["castbar_backdrop"]["a"] = 0.80
 		E.db["unitframe"]["colors"]["castbar_backdrop"]["b"] = 0.058823529411765
 		E.db["unitframe"]["colors"]["castbar_backdrop"]["g"] = 0.058823529411765
 		E.db["unitframe"]["colors"]["castbar_backdrop"]["r"] = 0.058823529411765
@@ -2042,5 +2045,5 @@ function L1UI:SetupLayout(layout)
 
 	E:StaggeredUpdateAll()
 
-	L1UI:Print('Layout has been set.')
+	L1UI:Print(L["Layout has been set."])
 end