+++ ElvUI_LuckyoneUI/Init.lua
@@ -1,11 +1,13 @@
-local E, L, V, P, G = unpack(ElvUI)
+local E, _, V, P, G = unpack(ElvUI)
+local L = E.Libs.ACL:GetLocale('ElvUI', E.global.general.locale)
 local EP = LibStub('LibElvUIPlugin-1.0')
 local addon, Engine = ...
+
+local format = format
 
 local _G = _G
 local GetAddOnMetadata = GetAddOnMetadata
 
--- Ace
 local L1UI = E:NewModule(addon, 'AceConsole-3.0', 'AceHook-3.0', 'AceEvent-3.0', 'AceTimer-3.0');
 
 Engine[1] = L1UI -- LuckyoneUI
@@ -20,25 +22,47 @@ _G[addon] = Engine;
 L1UI.Config = {}
 L1UI.CreditsList = {}
 L1UI.Name = '|cff4beb2cLuckyoneUI|r'
+L1UI.RequiredVersion = E.Retail and 12.64 or E.TBC and 2.36 or E.Classic and 1.62
 L1UI.Version = GetAddOnMetadata(addon, 'Version')
 
--- Load the following on login
-function L1UI:PLAYER_ENTERING_WORLD()
+local name = format('%s-%s', E.myname, E:ShortenRealm(E.myrealm))
+local toons = {
+	['Luckyone-LaughingSkull'] = true,
+	['Luckypriest-LaughingSkull'] = true,
+	['Luckymonkas-LaughingSkull'] = true,
+	['Luckydk-LaughingSkull'] = true,
+	['Luckyhunter-LaughingSkull'] = true,
+	['Unluckyone-LaughingSkull'] = true,
+	['Notlucky-LaughingSkull'] = true,
+	['Luckymage-LaughingSkull'] = true,
+	['Luckydh-LaughingSkull'] = true,
+	['Luckywl-LaughingSkull'] = true,
+	['Luckyrogue-LaughingSkull'] = true,
+	['Luckypala-LaughingSkull'] = true
+}
+
+function L1UI:PLAYER_ENTERING_WORLD(_, initLogin, isReload)
+	if initLogin or not ElvDB.LuckyoneDisabledAddOns then
+		ElvDB.LuckyoneDisabledAddOns = {}
+	end
+
+	if initLogin or isReload then L1UI:VersionCheck() end
+
 	L1UI:DisabledFrames()
 	L1UI:LoadCommands()
 end
 
--- Init
 function L1UI:Initialize()
 	if E.private.install_complete and E.private.L1UI.install_version == nil then
 		E.PluginInstaller:Queue(L1UI.InstallerData)
 	end
 
+	if toons[name] then ElvDB.ShadowLightAlpha = false end
+
 	EP:RegisterPlugin(addon, L1UI.GetOptions)
 	L1UI:RegisterEvent('PLAYER_ENTERING_WORLD')
 end
 
--- Callback
 local function CallbackInitialize()
 	L1UI:Initialize()
 end