+++ ElvUI_LuckyoneUI/Locales/enUS.lua
@@ -1,0 +1,179 @@
+local L = ElvUI[1].Libs.ACL:NewLocale('ElvUI', 'enUS')
+if not L then return end
+
+--[[
+	Please contact me on Discord if you contributed to the translation of LuckyoneUI.
+	I will verify your Pull Request on GitHub and add you to the AddOn credits.
+]]
+
+-- Config.lua locales
+L["Addon link:"] = true
+L["Addon Profiles"] = true
+L["Alert Frame"] = true
+L["Auras"] = true
+L["Author"] = true
+L["Available Tags"] = true
+L["Blizzard"] = true
+L["Boss Banner"] = true
+L["Boss Frame Buffs"] = true
+L["Boss Frame Debuffs"] = true
+L["BossMods Profiles"] = true
+L["Buffs"] = true
+L["Changelog:"] = true
+L["Chat CVars"] = true
+L["Chat"] = true
+L["Class Color Style"] = true
+L["Class Color"] = true
+L["Clear Cache"] = true
+L["Clear Chat History"] = true
+L["Clear Editbox History"] = true
+L["Coding"] = true
+L["Credits"] = true
+L["CVars"] = true
+L["Dark Style (Default)"] = true
+L["Dark"] = true
+L["Death Knight"] = true
+L["Debuffs"] = true
+L["Default Style"] = true
+L["Default"] = true
+L["Demon Hunter"] = true
+L["Description"] = true
+L["Disable Party Buffs"] = true
+L["DPS/TANK Layout"] = true
+L["Druid"] = true
+L["ElvUI Plugins"] = true
+L["Enable Party Buffs"] = true
+L["Focus Frame Buffs"] = true
+L["Focus Frame Debuffs"] = true
+L["Fonts, Textures & Skins"] = true
+L["General CVars"] = true
+L["General WeakAuras"] = true
+L["Healing Layout"] = true
+L["Hide Blizzard Frames"] = true
+L["Hide the Boss Banner"] = true
+L["Hide the Loot/Alert Frame"] = true
+L["Hide the Zone Text"] = true
+L["Hunter"] = true
+L["Install"] = true
+L["Installer & Update"] = true
+L["Jump to the Available Tag list."] = true
+L["Links"] = true
+L["Mage"] = true
+L["Media"] = true
+L["Minimalistic. Only important auras."] = true
+L["Minimalistic"] = true
+L["Monk"] = true
+L["NamePlate CVars"] = true
+L["NamePlate Profiles"] = true
+L["NamePlates: Enemy NPC Buffs"] = true
+L["NamePlates: Enemy NPC Debuffs"] = true
+L["Note"] = true
+L["Paladin"] = true
+L["Party Buffs"] = true
+L["Performance"] = true
+L["Player Frame Buffs"] = true
+L["Player Frame Debuffs"] = true
+L["Priest"] = true
+L["Profiles"] = true
+L["Raidframe Style"] = true
+L["Re-Run the installation process."] = true
+L["Report issues here:"] = true
+L["Reset Fonts, Textures, Skins to LuckyoneUI defaults."] = true
+L["Reset Media"] = true
+L["Reset to LuckyoneUI defaults."] = true
+L["Rogue"] = true
+L["Setup Chat will reset your chat panels to default and create custom chat tabs.\n\nLeft Chat: [ General - Log - Whisper - Guild - Party ]\n\nRight Chat: [ No Tabs - Details! Damage Meter ]"] = true
+L["Setup Chat"] = true
+L["Setup CVars"] = true
+L["Shaman"] = true
+L["Show all auras except blacklisted."] = true
+L["Show All"] = true
+L["Supporters"] = true
+L["Tags"] = true
+L["Target Frame Buffs"] = true
+L["Target Frame Debuffs"] = true
+L["Testers and Translation"] = true
+L["The options above are for the DPS/TANK layout."] = true
+L["Themes"] = true
+L["This will only clear AddOn Cache and not wipe your settings"] = true
+L["UnitFrames Color Theme"] = true
+L["UnitFrames"] = true
+L["Update DPS/TANK layout to LuckyoneUI version: "] = true
+L["Update DPS/TANK Layout"] = true
+L["Update Healing layout to LuckyoneUI version: "] = true
+L["Update Healing Layout"] = true
+L["Vertical Style"] = true
+L["Vertical"] = true
+L["Warlock"] = true
+L["Warrior"] = true
+L["WeakAuras DPS/TANK"] = true
+L["Wowhead Guide:"] = true
+L["Zone Text"] = true
+
+-- Core.lua locales
+L["AddOnSkins profile has been set."] = true
+L["BigWigs profile has been set."] = true
+L["CVars have been set."] = true
+L["DBM profile has been set."] = true
+L["Details profile has been set."] = true
+L["NamePlate CVars have been set."] = true
+L["OmniCD profile has been set."] = true
+L["Plater profile has been set."] = true
+L["ProjectAzilroka profile has been set."] = true
+L["Reload required - continue?"] = true
+L["Shadow&Light profile has been set."] = true
+L["Your ElvUI is outdated - please update and reload."] = true
+
+-- Install.lua locales
+L["BossMods"] = true
+L["Choose between DBM and BigWigs."] = true
+L["Choose between ElvUI NamePlates and Plater NamePlates."] = true
+L["Color Theme"] = true
+L["Details Profile"] = true
+L["DPS/Tank"] = true
+L["Finished"] = true
+L["Importance: |cff4beb2cHigh|r"] = true
+L["Importance: |cff4beb2cOptional|r"] = true
+L["Installation Complete"] = true
+L["Installation"] = true
+L["Layouts"] = true
+L["NamePlates"] = true
+L["Note: |cff4beb2cMake sure you have a backup of your current profile|r"] = true
+L["OmniCD Profile"] = true
+L["Please click a button below to apply the layout of your choosing."] = true
+L["Please click a button below to set the UnitFrames Color Theme."] = true
+L["Please click the button below to apply Luckyones profile for Details! Damage Meter."] = true
+L["Setup Details"] = true
+L["Setup OmniCD"] = true
+L["Skip Process"] = true
+L["This installation process will guide you through a few steps and apply settings to your current ElvUI profile."] = true
+L["This will setup the chat windows to look like this:\n\nLeft Chat: General - Log - Whisper - Guild - Party\n\nRight Chat: Details! Damage Meter"] = true
+L["WeakAuras Information"] = true
+L["Welcome"] = true
+L["You can find WeakAuras for all 12 Classes and some Utility WeakAuras in the LuckyoneUI config."] = true
+L["You have completed the installation process, please click 'Finished' to reload the UI."] = true
+
+-- Layout.lua locales
+L["Layout has been set."] = true
+
+-- Tags.lua locales
+L["Displays percentage health with 1 decimal below 100%, 2 decimals below 10% and hides decimals at 100%"] = true
+L["Displays the unit's classification (e.g 'Elite' and 'Rare') but without 'Affix'"] = true
+L["Displays the unit's Mana (Role: Healer)"] = true
+L["Displays the unit's Mana in percent (Role: Healer)"] = true
+
+-- Themes.lua locales
+L["Theme has been set."] = true
+
+-- Update.lua locales
+L["Successfully updated your current layout to LuckyoneUI version "] = true
+
+-- Auras.lua locales
+L["Buff filters updated."] = true
+L["Debuff filters updated."] = true
+
+-- Chat.lua locales
+L["Chat setup successful."] = true
+
+-- NamePlates.lua locales
+L["NamePlate StyleFilters and CVars have been set."] = true